[FP-129] refactor : Game Entity 필드 수정 및 위치 변경
