+++ game-service/src/main/java/com/fix/game_service/application/dtos/request/GameSearchRequest.java
@@ -3,7 +3,7 @@ package com.fix.game_service.application.dtos.request;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
-import com.fix.game_service.domain.Team;
+import com.fix.game_service.domain.model.Team;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
