+++ game-service/src/main/java/com/fix/game_service/application/dtos/response/GameGetOneResponse.java
@@ -3,9 +3,9 @@ package com.fix.game_service.application.dtos.response;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
-import com.fix.game_service.domain.Game;
-import com.fix.game_service.domain.GameStatus;
-import com.fix.game_service.domain.Team;
+import com.fix.game_service.domain.model.Game;
+import com.fix.game_service.domain.model.GameStatus;
+import com.fix.game_service.domain.model.Team;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -20,8 +20,8 @@ public class GameGetOneResponse {
 
 	private UUID gameId;
 	private String gameName;
-	private Team gameTeam1;
-	private Team gameTeam2;
+	private Team homeTeam;
+	private Team awayTeam;
 	private LocalDateTime gameDate;
 	private UUID stadiumId;
 	private GameStatus gameStatus;
@@ -32,8 +32,8 @@ public class GameGetOneResponse {
 		return GameGetOneResponse.builder()
 			.gameId(game.getGameId())
 			.gameName(game.getGameName())
-			.gameTeam1(game.getGameTeam1())
-			.gameTeam2(game.getGameTeam2())
+			.homeTeam(game.getHomeTeam())
+			.awayTeam(game.getAwayTeam())
 			.gameDate(game.getGameDate())
 			.stadiumId(game.getStadiumId())
 			.gameStatus(game.getGameStatus())
