+++ game-service/src/main/java/com/fix/game_service/domain/model/Team.java
@@ -1,4 +1,4 @@
-package com.fix.game_service.domain;
+package com.fix.game_service.domain.model;
 
 import lombok.Getter;
